pyproject.toml
[project]
name = "pubmed-fetcher"
version = "0.1.0"
description = ""
authors = [
    {name = "Grandhi Ganga Phanindra Kumar",email = "phanigrandhi91@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "typer (>=0.15.2,<0.16.0)",
    "biopython (>=1.85,<2.0)",
    "flask (>=3.1.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
